Updated in case of no payment_id in misc_credit
Added code to solve the issue when there is no payment in a particular payment_id.
Ex -I recieved payment from x in form of cash, but I did not pay anything in form of cash. So in this case it wont be joined by previous SP code. Now this issue will be resolved.